Break Linux::Writer into smaller classes (API to be revisited)
